+++ my_project/dao/river_dao.py
@@ -1,13 +1,18 @@
 from extensions import db
 from my_project.domain.river import River
 from sqlalchemy import select
+from sqlalchemy.exc import IntegrityError
 
 
 class RiverDao:
     def create(self, name, length):
         r = River(name=name, length=length)  # type: ignore
         db.session.add(r)
-        db.session.commit()
+        try:
+            db.session.commit()
+        except IntegrityError:
+            db.session.rollback()
+            return None
         return r
 
     def get(self, id):
@@ -20,12 +25,25 @@ class RiverDao:
         ).mappings().all()
         return list
 
-    def delete(self, r: River):
+    def delete(self, id):
+        r = self.get(id)
+        if r is None:
+            return False
         db.session.delete(r)
         db.session.commit()
+        return True
 
-    def update(self, r, new_name, new_length):
+    def update(self, id, new_name, new_length):
+        r = self.get(id)
+        if r is None:
+            return (False, None)
+
         r.name = new_name
         r.length = new_length
-        db.session.commit()
-        return r
+
+        try:
+            db.session.commit()
+        except IntegrityError:
+            db.session.rollback()
+            return (True, None)
+        return (True, r)
